# artifacts/api-server/src/routes/leads.ts
import { Router, type IRouter } from "express";
import { eq, ilike, and, type SQL } from "drizzle-orm";
import { db, leadsTable, agentsTable, activityTable } from "@workspace/db";
import {
  CreateLeadBody,
  UpdateLeadBody,
  UpdateLeadParams,
  GetLeadParams,
} from "@workspace/api-zod";

const router: IRouter = Router();

function formatLead(lead: typeof leadsTable.$inferSelect, agentName?: string | null) {
  return {
    ...lead,
    budget: lead.budget != null ? Number(lead.budget) : null,
    agentName: agentName ?? null,
    createdAt: lead.createdAt.toISOString(),
    updatedAt: lead.updatedAt.toISOString(),
  };
}

router.get("/leads", async (req, res): Promise<void> => {
  const { status, search, assignedTo } = req.query;
  const conditions: SQL[] = [];
  if (status && typeof status === "string") conditions.push(eq(leadsTable.status, status));
  if (assignedTo) conditions.push(eq(leadsTable.assignedTo, Number(assignedTo)));

  let rows;
  if (conditions.length > 0) {
    rows = await db.select().from(leadsTable).where(and(...conditions)).orderBy(leadsTable.createdAt);
  } else {
    rows = await db.select().from(leadsTable).orderBy(leadsTable.createdAt);
  }

  // Filter by search if provided
  let filtered = rows;
  if (search && typeof search === "string") {
    const lower = search.toLowerCase();
    filtered = rows.filter(
      (r) => r.name.toLowerCase().includes(lower) || r.email.toLowerCase().includes(lower),
    );
  }

  const agents = await db.select().from(agentsTable);
  const agentMap = new Map(agents.map((a) => [a.id, a.name]));

  res.json(filtered.map((l) => formatLead(l, agentMap.get(l.assignedTo ?? -1))));
});

router.post("/leads", async (req, res): Promise<void> => {
  const parsed = CreateLeadBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }
  const [lead] = await db.insert(leadsTable).values({ ...parsed.data, status: parsed.data.status ?? "new", score: parsed.data.score ?? 50 }).returning();

  let agentName: string | null = null;
  if (lead.assignedTo) {
    const [agent] = await db.select().from(agentsTable).where(eq(agentsTable.id, lead.assignedTo));
    agentName = agent?.name ?? null;
  }

  await db.insert(activityTable).values({
    type: "lead_created",
    description: `New lead created`,
    entityName: lead.name,
    agentId: lead.assignedTo ?? undefined,
  });

  res.status(201).json(formatLead(lead, agentName));
});

router.get("/leads/:id", async (req, res): Promise<void> => {
  const raw = Array.isArray(req.params.id) ? req.params.id[0] : req.params.id;
  const params = GetLeadParams.safeParse({ id: Number(raw) });
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }
  const [lead] = await db.select().from(leadsTable).where(eq(leadsTable.id, params.data.id));
  if (!lead) {
    res.status(404).json({ error: "Lead not found" });
    return;
  }
  let agentName: string | null = null;
  if (lead.assignedTo) {
    const [agent] = await db.select().from(agentsTable).where(eq(agentsTable.id, lead.assignedTo));
    agentName = agent?.name ?? null;
  }
  res.json(formatLead(lead, agentName));
});

router.put("/leads/:id", async (req, res): Promise<void> => {
  const raw = Array.isArray(req.params.id) ? req.params.id[0] : req.params.id;
  const params = UpdateLeadParams.safeParse({ id: Number(raw) });
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }
  const parsed = UpdateLeadBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }
  const [lead] = await db
    .update(leadsTable)
    .set(parsed.data)
    .where(eq(leadsTable.id, params.data.id))
    .returning();
  if (!lead) {
    res.status(404).json({ error: "Lead not found" });
    return;
  }
  let agentName: string | null = null;
  if (lead.assignedTo) {
    const [agent] = await db.select().from(agentsTable).where(eq(agentsTable.id, lead.assignedTo));
    agentName = agent?.name ?? null;
  }
  res.json(formatLead(lead, agentName));
});

router.delete("/leads/:id", async (req, res): Promise<void> => {
  const raw = Array.isArray(req.params.id) ? req.params.id[0] : req.params.id;
  const id = parseInt(raw, 10);
  if (isNaN(id)) {
    res.status(400).json({ error: "Invalid id" });
    return;
  }
  const [lead] = await db.delete(leadsTable).where(eq(leadsTable.id, id)).returning();
  if (!lead) {
    res.status(404).json({ error: "Lead not found" });
    return;
  }
  res.sendStatus(204);
});

export default router;