# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EstateFlow CRM API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Lead {
  id: number;
  name: string;
  email: string;
  phone?: string | null;
  /** zillow, realtor, referral, website, phone, email, other */
  source: string;
  /** new, contacted, qualified, proposal, negotiation, closed_won, closed_lost */
  status: string;
  /** 0-100 lead score */
  score: number;
  budget?: number | null;
  /** residential, commercial, land, rental */
  propertyType?: string | null;
  notes?: string | null;
  /** Agent ID */
  assignedTo?: number | null;
  agentName?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateLeadBody {
  name: string;
  email: string;
  phone?: string | null;
  source: string;
  status?: string;
  score?: number;
  budget?: number | null;
  propertyType?: string | null;
  notes?: string | null;
  assignedTo?: number | null;
}

export interface UpdateLeadBody {
  name?: string;
  email?: string;
  phone?: string | null;
  source?: string;
  status?: string;
  score?: number;
  budget?: number | null;
  propertyType?: string | null;
  notes?: string | null;
  assignedTo?: number | null;
}

export interface Property {
  id: number;
  title: string;
  address: string;
  city: string;
  price: number;
  /** residential, commercial, land, rental */
  type: string;
  /** available, under_offer, sold, withdrawn */
  status: string;
  bedrooms?: number | null;
  bathrooms?: number | null;
  areaSqft?: number | null;
  description?: string | null;
  agentId?: number | null;
  agentName?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreatePropertyBody {
  title: string;
  address: string;
  city: string;
  price: number;
  type: string;
  status?: string;
  bedrooms?: number | null;
  bathrooms?: number | null;
  areaSqft?: number | null;
  description?: string | null;
  agentId?: number | null;
}

export interface UpdatePropertyBody {
  title?: string;
  address?: string;
  city?: string;
  price?: number;
  type?: string;
  status?: string;
  bedrooms?: number | null;
  bathrooms?: number | null;
  areaSqft?: number | null;
  description?: string | null;
  agentId?: number | null;
}

export interface Deal {
  id: number;
  title: string;
  leadId?: number | null;
  leadName?: string | null;
  propertyId?: number | null;
  propertyTitle?: string | null;
  agentId?: number | null;
  agentName?: string | null;
  /** prospect, viewing, offer, under_contract, due_diligence, closed_won, closed_lost */
  stage: string;
  value: number;
  closingDate?: string | null;
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateDealBody {
  title: string;
  leadId?: number | null;
  propertyId?: number | null;
  agentId?: number | null;
  stage?: string;
  value: number;
  closingDate?: string | null;
  notes?: string | null;
}

export interface UpdateDealBody {
  title?: string;
  leadId?: number | null;
  propertyId?: number | null;
  agentId?: number | null;
  stage?: string;
  value?: number;
  closingDate?: string | null;
  notes?: string | null;
}

export interface Agent {
  id: number;
  name: string;
  email: string;
  phone?: string | null;
  /** admin, manager, agent, support */
  role: string;
  /** US, UK, AU, NZ, EU */
  market?: string | null;
  activeLeads: number;
  dealsCount: number;
  revenue: number;
  createdAt: string;
}

export interface CreateAgentBody {
  name: string;
  email: string;
  phone?: string | null;
  role?: string;
  market?: string | null;
}

export interface UpdateAgentBody {
  name?: string;
  email?: string;
  phone?: string | null;
  role?: string;
  market?: string | null;
}

export interface DashboardStats {
  totalLeads: number;
  newLeadsThisMonth: number;
  totalDeals: number;
  activeDeals: number;
  pipelineValue: number;
  closedRevenue: number;
  totalProperties: number;
  availableProperties: number;
  conversionRate: number;
  avgDealValue: number;
}

export interface PipelineStage {
  stage: string;
  count: number;
  value: number;
}

export interface ActivityItem {
  id: number;
  /** lead_created, deal_updated, property_added, deal_closed */
  type: string;
  description: string;
  entityName: string;
  agentName?: string | null;
  createdAt: string;
}

export interface LeadSource {
  source: string;
  count: number;
}

export type GetLeadsParams = {
  status?: string;
  search?: string;
  assignedTo?: number;
};

export type GetPropertiesParams = {
  status?: string;
  type?: string;
};

export type GetDealsParams = {
  stage?: string;
  agentId?: number;
};