# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EstateFlow CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leads:
    get:
      operationId: getLeads
      tags: [leads]
      summary: List all leads
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: assignedTo
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create a lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeadBody"
      responses:
        "201":
          description: Created lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get a lead by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Not found
    put:
      operationId: updateLead
      tags: [leads]
      summary: Update a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLeadBody"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /deals:
    get:
      operationId: getDeals
      tags: [deals]
      summary: List all deals
      parameters:
        - name: stage
          in: query
          schema:
            type: string
        - name: agentId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of deals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deal"
    post:
      operationId: createDeal
      tags: [deals]
      summary: Create a deal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDealBody"
      responses:
        "201":
          description: Created deal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
  /deals/{id}:
    get:
      operationId: getDeal
      tags: [deals]
      summary: Get a deal by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deal detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
        "404":
          description: Not found
    put:
      operationId: updateDeal
      tags: [deals]
      summary: Update a deal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDealBody"
      responses:
        "200":
          description: Updated deal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
    delete:
      operationId: deleteDeal
      tags: [deals]
      summary: Delete a deal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /agents:
    get:
      operationId: getAgents
      tags: [agents]
      summary: List all agents
      responses:
        "200":
          description: List of agents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Agent"
    post:
      operationId: createAgent
      tags: [agents]
      summary: Create an agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgentBody"
      responses:
        "201":
          description: Created agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
  /agents/{id}:
    get:
      operationId: getAgent
      tags: [agents]
      summary: Get an agent by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Agent detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "404":
          description: Not found
    put:
      operationId: updateAgent
      tags: [agents]
      summary: Update an agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAgentBody"
      responses:
        "200":
          description: Updated agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
...
components:
  schemas:
...
    Lead:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        source:
          type: string
          description: "zillow, realtor, referral, website, phone, email, other"
        status:
          type: string
          description: "new, contacted, qualified, proposal, negotiation, closed_won, closed_lost"
        score:
          type: integer
          description: 0-100 lead score
        budget:
          type: number
          nullable: true
        propertyType:
          type: string
          nullable: true
          description: "residential, commercial, land, rental"
        notes:
          type: string
          nullable: true
        assignedTo:
          type: integer
          nullable: true
          description: Agent ID
        agentName:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, email, source, status, score, createdAt, updatedAt]
    CreateLeadBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        source:
          type: string
        status:
          type: string
        score:
          type: integer
        budget:
          type: number
          nullable: true
        propertyType:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        assignedTo:
          type: integer
          nullable: true
      required: [name, email, source]
    UpdateLeadBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        source:
          type: string
        status:
          type: string
        score:
          type: integer
        budget:
          type: number
          nullable: true
        propertyType:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        assignedTo:
          type: integer
          nullable: true
...
    Deal:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        leadId:
          type: integer
          nullable: true
        leadName:
          type: string
          nullable: true
        propertyId:
          type: integer
          nullable: true
        propertyTitle:
          type: string
          nullable: true
        agentId:
          type: integer
          nullable: true
        agentName:
          type: string
          nullable: true
        stage:
          type: string
          description: "prospect, viewing, offer, under_contract, due_diligence, closed_won, closed_lost"
        value:
          type: number
        closingDate:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, stage, value, createdAt, updatedAt]
    CreateDealBody:
      type: object
      properties:
        title:
          type: string
        leadId:
          type: integer
          nullable: true
        propertyId:
          type: integer
          nullable: true
        agentId:
          type: integer
          nullable: true
        stage:
          type: string
        value:
          type: number
        closingDate:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required: [title, value]
    UpdateDealBody:
      type: object
      properties:
        title:
          type: string
        leadId:
          type: integer
          nullable: true
        propertyId:
          type: integer
          nullable: true
        agentId:
          type: integer
          nullable: true
        stage:
          type: string
        value:
          type: number
        closingDate:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
    Agent:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        role:
          type: string
          description: "admin, manager, agent, support"
        market:
          type: string
          nullable: true
          description: "US, UK, AU, NZ, EU"
        activeLeads:
          type: integer
        dealsCount:
          type: integer
        revenue:
          type: number
        createdAt:
          type: string
      required: [id, name, email, role, activeLeads, dealsCount, revenue, createdAt]
    CreateAgentBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        role:
          type: string
        market:
          type: string
          nullable: true
      required: [name, email]
    UpdateAgentBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        role:
          type: string
        market:
          type: string
          nullable: true
...